renamed and moved tests in test_creation.py and test_properties.py
